Colocate actions in a source file and clarify power actions
