+++ config/crd/bases/bmc.tinkerbell.org_tasks.yaml
@@ -90,8 +90,7 @@ spec:
                           type: string
                         type: array
                       efiBoot:
-                        description: EFIBoot specifies to EFI boot for the baseboard
-                          management.
+                        description: EFIBoot instructs the machine to use EFI boot.
                         type: boolean
                     required:
                     - device
